requirements.txt
cherrypy==18.9.0
google-generativeai==0.3.2
python-dotenv==1.0.0
